requirements.txt
transformers==4.41.2
torch==2.3.0
accelerate==0.30.1
Pillow==10.3.0
streamlit==1.35.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
sentencepiece==0.2.0
ollama==0.2.1
sentence-transformers==2.7.0
